Created a carousel component that auto fits the pictures to the screen based on width of the screen
This here is the "pièce de résistance" of the whole project. with the help of the custom usewindowsize hook dynamically resizes the shown pictures so they fit nicely in a grid and resize automagically when you change the widow size (even during the change didn't add debauncer because it is : a cool as heck and b: no one constantly resizes their window to worry about performance so much.